Ownership soundness: sanitizer, runtime checks, six holes, audit grid
Sanitizer and runtime:
- --san marks every generated function sanitize_address, so AddressSanitizer
  instruments reads and writes (it previously checked only malloc/free), and
  --san now also applies with -do. System tests compile with --san.
- Integer division by zero and INT64_MIN / -1 are runtime errors
  (_top_division_error) instead of undefined sdiv.
- Number literals are 64-bit; out-of-range literals are parse errors instead of
  an uncaught std::out_of_range.

Ownership (each previously accepted and memory-unsafe or leaking):
- A generic body may not use a formal after passing it on (Self copy
  requirement, judged at the call; requirements now record their reasons).
- Moves take effect in evaluation order, including writes through a moved
  owning pointer; a call may not move an owner one of its actuals borrows.
- A while condition may not move; a loop body may not change ownership.
- Joins: a variable owned on one path must be owned on every path
  (MoveAnalysis::joinStates, shared by DestructionPass).
- Owned binders of a by-value case are freed at the end of their arm.
- An owning reference produced by a call or alloc and dereferenced directly is
  freed after the read or write.
- Case binders are scoped to their arm: reused binder names no longer leak or
  let an alias escape AliasCheck; codegen restores shadowed binders.
- Diagnostics name the rule for loop-local owners and generic overwrites.

Audit:
- test/system/soundness: a generated grid of 1,118 programs (owner kind x
  operation x context) with verdicts derived from the ownership rules, run
  explicitly with run.py --soundness; adequacy.py disables each rule
  (--unsafe-disable / TOPC_UNSAFE_DISABLE) and requires a test to fail.
- The indirect-call "targets disagree" check is unreachable and is now an
  InternalError.

Tests: 13 new unit tests, 25 rejection iotests, a division iotest, 9 selftests;
two existing tests updated (moveInBothBranches leaked; generic write-through
message) and one ownership snapshot gained two trace events.

+++ src/frontend/ast/treetypes/ASTNumberExpr.h
@@ -2,14 +2,16 @@
 
 #include "ASTExpr.h"
 
+#include <cstdint>
+
 /*! \brief Class for numeric literals.
  */
 class ASTNumberExpr : public ASTExpr {
-  int VAL;
+  int64_t VAL;
 
 public:
-  ASTNumberExpr(int VAL) : VAL(VAL) {}
-  int getValue() const { return VAL; }
+  ASTNumberExpr(int64_t VAL) : VAL(VAL) {}
+  int64_t getValue() const { return VAL; }
   void accept(ASTVisitor *visitor) override;
 
 protected:
